% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{B_B.H}
\alias{B_B.H}
\title{Compare the effect of Bonferroni and B-H method.}
\usage{
B_B.H(M)
}
\arguments{
\item{M}{the number of simulations.}
}
\value{
the results of the two methods in terms of FWER, FDR and TPR.
}
\description{
the evaluating metrics are FWER, FDR and TPR.
}
\examples{
\dontrun{
    B_B.H(1000)
}
}
