% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{em}
\alias{em}
\title{An EM algorithm to estimate the lambda of exponential distribution.}
\usage{
em(x0, u, v, max.it = 500, eps = 1e-07)
}
\arguments{
\item{x0}{initial value for the iterative procedure.}

\item{u, v}{the known interval.}

\item{max.it}{the maximum iterative times, the default is 500.}

\item{eps}{the threshold condition for the end of the iteration, the default is 1e-7.}
}
\value{
an estimator for lambda by EM algorithm.
}
\description{
An EM algorithm to estimate the lambda of exponential distribution.
}
\examples{
\dontrun{
    u<-c(11,8,27,13,16,0,23,10,24,2)
    v<-c(12,9,28,14,17,1,24,11,25,3)
    em(1,u,v)
}
}
