% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{Gelman.Rubin}
\alias{Gelman.Rubin}
\title{Gelman-Rubin method to monitor convergence of the chain}
\usage{
Gelman.Rubin(psi)
}
\arguments{
\item{psi}{psi[i,j] is the statistic psi(X[i,1:j]) for chain in i-th row of X.}
}
\value{
hat_r
}
\description{
Refer to Example 9.1. Use the Gelman-Rubin method to monitor convergence of the chain, and run the chain until the chain has converged approximately to the target distribution.
}
