% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{jackn}
\alias{jackn}
\title{Jackknife estimate}
\usage{
jackn(data)
}
\arguments{
\item{data}{a dataframe.}
}
\value{
an vector of length 2 including the jackknife estimates of bias and standard error.
}
\description{
Refer to Exercise 7.7. Obtain the jackknife estimates of bias and standard error of theta
}
\examples{
\dontrun{
    attach(scor)
    jackn9(scor)
    detach(scor)
}
}
