% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{mc}
\alias{mc}
\title{Use MC to estimate integration by simple MC and antithetic variate approach}
\usage{
mc(K)
}
\arguments{
\item{K}{the number of repeated trials.}
}
\value{
an list includes the results by the antithetic bariate approach and by the simple Monte Carlo method under \code{K} times.
}
\description{
Use the antithetic variate approach and the simple Monte Carlo method to estimate the value of theta.
}
\examples{
\dontrun{
    result<-mc(1000)
    mean(result$simple)
    mean(result$anti)
    (var(result$simple)-var(result$anti))/var(result$simple)
}
}
