% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{mc1}
\alias{mc1}
\title{Use MC to estimate integration by importance sampling}
\usage{
mc1(n)
}
\arguments{
\item{n}{the number of sampling.}
}
\value{
an list includes mean and est by two different importance funcion.
}
\description{
Use the importance sampling method to estimate the value of theta.
}
\examples{
\dontrun{
    mc1(1e6)
}
}
