% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{mc2}
\alias{mc2}
\title{Use MC to estimate integration by the stratified importance sampling}
\usage{
mc2(n)
}
\arguments{
\item{n}{the number of sampling.}
}
\value{
an list includes mean and est by the stratified importance sampling.
}
\description{
Use the stratified importance sampling method to estimate the value of theta.
}
\examples{
\dontrun{
    mc2(2000)
}
}
