% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{mc3}
\alias{mc3}
\title{Test the empirical Type 1 error under MC estimation.}
\usage{
mc3(n, k)
}
\arguments{
\item{n}{the number of random numbers.}

\item{k}{the number of trials.}
}
\value{
an list includes empirical Type 1 error under three different distribution.
}
\description{
according to E6.A.
}
\examples{
\dontrun{
    mc3(20,1e4)
}
}
