% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{my.boots}
\alias{my.boots}
\title{My bootstrap function to estimate the parameter of exponential distribution.}
\usage{
my.boots(n)
}
\arguments{
\item{n}{the size of sample.}
}
\value{
compare the mean bootstrap bias and bootstrap standard error woth the thepretical ones.
}
\description{
My bootstrap function to estimate the parameter of exponential distribution.
}
\examples{
\dontrun{
    set.seed(1234)
    n5<-boots(5)
    n10<-boots(10)
    n20<-boots(20)
    round(rbind(n5,n10,n20),4)
}
}
