% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{my.gibbs}
\alias{my.gibbs}
\title{Gibbs sampler to generate a bivariate normal chain}
\usage{
my.gibbs(N, burn)
}
\arguments{
\item{N}{the length of chain.}

\item{burn}{burn-in length.}
}
\value{
a random vector with the burn-in period removed.
}
\description{
Implement a Gibbs sampler to generate a bivariate normal chain with zero means, unit standard deviations, and correlation 0.9.
}
\examples{
\dontrun{
    my.gibbs(15000,5000)
}
}
