% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{my.laplace}
\alias{my.laplace}
\title{My Laplace function}
\usage{
my.laplace(n)
}
\arguments{
\item{n}{the number of samples.}
}
\value{
a random sample of size \code{n}
}
\description{
Use the inverse transform method to generate a random sample from Laplace distribution
}
\examples{
\dontrun{
    X<-my.laplace(1000)
}
}
