% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{my.sample}
\alias{my.sample}
\title{My sample function}
\usage{
my.sample(x, size, prob = NULL)
}
\arguments{
\item{x}{the data.}

\item{size}{the number of samples.}

\item{prob}{the sample probability given to the data \code{x}.}
}
\value{
a random sample of size \code{size}
}
\description{
Use the inverse transform method to realize sample function when replace equals True.
}
\examples{
\dontrun{
    n<-1000
    X1<-my.sample(1:5,n)
    X2<-my.sample(1:5,n,prob = c(1,2,3,4,5))
    X3<-my.sample(c(3,2,1,4,5),n,prob = c(5,4,3,2,1))
}
}
