% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{my_pi}
\alias{my_pi}
\title{Estimate the value of pi}
\usage{
my_pi(rho)
}
\arguments{
\item{rho}{the ratio of l to d, shoud be a numeric between 0 and 1.}
}
\value{
an estimated value of pi.
}
\description{
Use needle dropping methods to estimate the value of pi.
}
\examples{
\dontrun{
    my_pi(0.5)
    my_pi(0.8)
    my_pi(1)
}
}
