% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{r_gibbs}
\alias{r_gibbs}
\title{The Gibbs sampler for the binomial-beta distribution.}
\usage{
r_gibbs(num, a, b, n)
}
\arguments{
\item{num}{number of observations.}

\item{a, b, n}{the parameters of the density, keep a close watch on the description for details.}
}
\value{
a vector of target random numbers.
}
\description{
The target bivariate density can be shown that for fixed a,b,n, the conditional distributions are B(n,y) and Beta(x+a,n−x+b).
}
\examples{
\dontrun{
    r_gibbs(1000,2,3,10)
}
}
