% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_gibbs}
\alias{rcpp_gibbs}
\title{A Gibbs sampler using Rcpp}
\usage{
rcpp_gibbs(N, a, b, n)
}
\arguments{
\item{N}{the number of samples}

\item{a, b, n}{the parameters of the target bivariate density.}
}
\value{
a random sample of size \code{N}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
 rcpp_gibbs(1000,2,3,10)
 }
}
