% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{rEpa}
\alias{rEpa}
\title{My random rescaled Epanechnikov kernel function}
\usage{
rEpa(n)
}
\arguments{
\item{n}{the number of samples.}
}
\value{
a random sample of size \code{n}
}
\description{
Use the algorithm in E3.9 to generate a random sample from the rescaled Epanechnikov kernel.
}
\examples{
\dontrun{
    X<-rEpa(1000)
}
}
