% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{Random walk Metropolis sampler for generating the standard Laplace distribution}
\usage{
rw.Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{standard variance of proposal distribution N(x_t,sigma).}

\item{x0}{initial value.}

\item{N}{size of random numbers required.}
}
\value{
a list including a vector of random numbers and repetition times.
}
\description{
a function that generates the standard Laplace distribution random numbers by random walk Metropolis sampling.
}
\examples{
\dontrun{
    rw.Metropolis(.5, 100, 10000)
}
}
