\name{SA23204179-package}
\alias{SA23204179-package}
\alias{SA23204179}
\docType{package}
\title{
  Example functions for 'Statistical Computing' course
}
\description{
 Several R and Rcpp function are included in this package for illustration. 
}
\details{
 There are two kinds funcions, all my homework answers in statistical computing 2023fall and two funcions related to my research.
}
\author{
Jian Yang

Maintainer: Jian Yang <1326662908@qq.com> 
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }

