% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{solve_alpha}
\alias{solve_alpha}
\title{Estimate a in Logistic model}
\usage{
solve_alpha(N, b1, b2, b3, f0)
}
\arguments{
\item{N}{the number of random numbers.}

\item{b1, b2, b3}{the parameters in logistic model.}

\item{f0}{a parameter.}
}
\value{
a numeric.
}
\description{
a function that takes as input values N,b1,b2,b3 and f0, and produces the output a.
}
\examples{
\dontrun{
    solve_alpha(1e6,0,1,-1,0.1)
}
}
