% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{W2}
\alias{W2}
\title{Compute W-2 statistics}
\usage{
W2(z, i, sizes)
}
\arguments{
\item{z}{data}

\item{i}{index, a vector.}

\item{sizes}{sizes.}
}
\value{
a numeric.
}
\description{
Implement the two-sample Cram`er-von Mises test for equal distributions as a permutation test.
}
\examples{
\dontrun{
    attach(chickwts)
    x <- as.vector(weight[feed == "soybean"]);y <- as.vector(weight[feed == "linseed"])
    detach(chickwts)
    z <- c(x, y)
    N <- c(length(x), length(y))
    boot.obj <- boot(data = z, statistic = W2, R = 9999,sim = "permutation", sizes = N)
    ts <- c(boot.obj$t0,boot.obj$t)
    p.value <- mean(ts>=ts[1])
}
}
